Initial commit
Summary:
Budgie panel applet to manage the user's trash bin.
Allows the user to restore or delete individual or all trash items from the Budgie panel.

I was told by JoshStrobl to skip the package inclusion step for this, which is why none exists for this.

Test Plan:
Build and install, verify the applet can be added to a Budgie panel.
Verify that the individual item restore/delete functions work correctly.
Verify that the restore/delete all functions work correctly.

Reviewers: #triage_team, Girtablulu, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl, Harvey, Girtablulu

Differential Revision: https://dev.getsol.us/D6070